First pass swipe extraction algorithm
This commit adds the first pass implementation of a swipe extraction
algorithm. The algorithm is implemented by first extracting the unique
sentences the participants swiped against. Then for each unique
sentence string, the string is put into a dictionary as the key, and
stored with a list of values corresponding the remainder of the string
for that particular occurrence of the sentence key. This allows us to
store the sentence once while still keeping its association with the
rest of its corresponding data for a particular row in the log file.

The next step is to abstract an individual value list as a "Swipe"
data structure where there will be getters for all of the column values
for that particular "swipe". Then we can put all of these swipes into an
array which stores the sentence key and an array of the swipes, calling it
then a "Swipe Set".