First pass implementation for swipe extraction across many words.
This commit implements new functions to make swipe extraction work with
many words instead of a single word. The current workflow is to take
a dataset file and extract the unique words present in the file and
their associated trajectories and save them to their own files. Then
the rest of the protocol works the same. During this process, I found
a few edge cases that I noted down in edge_cases.py. There were 2
categories of edge cases:
    1) We need to gracefully handle the situation when none of the time
    deltas are above the threshold, so the rest of the pipeline should
    not run.
    2) There were several instances when the returned indices were [0,0]
    We need to handle this situation somehow but I'm not sure exactly
    how we want to do that.
Next steps would be to apply the same code to all the database files and
see the results and also handle the edge cases.