Add a backing file for each swipe
This commit adds a new Backing_File data structure to each swipe. Its
purpose is to help encapsulate the rest of the data associated with a
swipe. The Backing_File stores the file path for the file that the swipe
was created from. It then exposes a lookup API to return a particular
row of the file based on the timestamp. So by providing the timestamp
it performs a lookup for the particular row of the file with the
matching timestamp. A next step for this would be that if multiple
timestamps match return all the associated rows. Now we should be able
to combine this with yet to be implemented helpers to extract features