Scaffold a class for genuine/imposter user
This commit scaffolds a new class that will come to represent genuine and imposter users. These "users" take in a file from the dataset as input and store created swipes.
These swipes can then be split in a 70:30 ratio to create the distinct genuine and imposter users. From there extract the feature vectors and calculate the manhattan distance