Drastically improve swipe extraction performance
This commit improves the performance of swipe extraction to around 80
seconds. But we still need to triple check that we did not regress in
accuracy of the swipes.
There is also a new method to write all the trajectories for every
dataset file, but this is very slow and should only be run once
(hopefully)